refactor: rename service package to application
