+++ src/main/java/com/example/minimart/order/controller/OrderController.java
@@ -2,7 +2,7 @@ package com.example.minimart.order.controller;
 
 import com.example.minimart.order.controller.dto.request.CreateOrderRequest;
 import com.example.minimart.order.infra.entity.OrderEntity;
-import com.example.minimart.order.service.OrderService;
+import com.example.minimart.order.application.OrderService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.tags.Tag;
